requirements.txt
pandas
numpy
matplotlib
yfinance
nltk
torch
scikit-learn
streamlit
newsapi-python
requests
python-dotenv
joblib
beautifulsoup4
